# src/DemoApiForGeneratingStuff/DemoApiForGeneratingStuff/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IssueTracker.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/catalog/{id}": {
      "delete": {
        "tags": [
          "Software Catalog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Software Catalog"
        ],
        "operationId": "Replace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCatalogItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCatalogItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCatalogItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Software Catalog"
        ],
        "operationId": "catalog#get-by-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CatalogItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReplaceCatalogItemRequest": {
        "required": [
          "description",
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "maxLength": 256,
            "minLength": 5,
            "type": "string"
          },
          "description": {
            "maxLength": 1024,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header with bearer token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}